Move stealSkin() into a new class
