+++ Wurst Client/src/tk/wurst_client/gui/alts/GuiAltLogin.java
@@ -23,6 +23,7 @@ import org.lwjgl.opengl.GL11;
 
 import tk.wurst_client.Client;
 import tk.wurst_client.alts.AltUtils;
+import tk.wurst_client.alts.SkinStealer;
 import tk.wurst_client.utils.MiscUtils;
 
 public class GuiAltLogin extends GuiScreen
@@ -104,7 +105,7 @@ public class GuiAltLogin extends GuiScreen
 			}else if(clickedButton.id == 3)
 				nameBox.setText(AltUtils.generateName());
 			else if(clickedButton.id == 4)
-				displayText = AltUtils.stealSkin(nameBox.getText());
+				displayText = SkinStealer.stealSkin(nameBox.getText());
 			else if(clickedButton.id == 5)
 				MiscUtils.openFile(Client.wurst.fileManager.skinDir);
 	}
